move accounts-general to methods
